Extract LeftRightMatrix to separate module
